Run prettier

+++ contracts/NitroAdjudicator.sol
@@ -4,55 +4,60 @@ import "openzeppelin-solidity/contracts/math/SafeMath.sol";
 import "./Commitment.sol";
 import "./Rules.sol";
 
-contract IERC20 { // Abstraction of the parts of the ERC20 Interface that we need
-    function transfer(address to, uint tokens) public returns (bool success);
-    function transferFrom(address from, address to, uint tokens) public returns (bool success);
+contract IERC20 {
+    // Abstraction of the parts of the ERC20 Interface that we need
+    function transfer(address to, uint256 tokens) public returns (bool success);
+    function transferFrom(address from, address to, uint256 tokens) public returns (bool success);
 }
 
-contract INitroLibrary { // Abstraction of the NitroLibrary contract
+contract INitroLibrary {
+    // Abstraction of the NitroLibrary contract
 
     struct Outcome {
-    address[] destination;
-    uint256 finalizedAt;
-    Commitment.CommitmentStruct challengeCommitment;
-    // exactly one of the following two should be non-null
-    // guarantee channels
-    uint[] allocation;         // should be zero length in guarantee channels
-    address[] token;
+        address[] destination;
+        uint256 finalizedAt;
+        Commitment.CommitmentStruct challengeCommitment;
+        // exactly one of the following two should be non-null
+        // guarantee channels
+        uint256[] allocation; // should be zero length in guarantee channels
+        address[] token;
     }
 
     struct Signature {
-    uint8 v;
-    bytes32 r;
-    bytes32 s;
+        uint8 v;
+        bytes32 r;
+        bytes32 s;
     }
 
-    function recoverSigner(bytes calldata _d, uint8 _v, bytes32 _r, bytes32 _s) external pure returns(address);
+    function recoverSigner(bytes calldata _d, uint8 _v, bytes32 _r, bytes32 _s)
+        external
+        pure
+        returns (address);
 
-    function affords(
-        address recipient,
-        Outcome calldata outcome,
-        uint funding
-    ) external pure returns (uint256);
+    function affords(address recipient, Outcome calldata outcome, uint256 funding)
+        external
+        pure
+        returns (uint256);
 
-    function reprioritize(
-        Outcome calldata allocation,
-        Outcome calldata guarantee
-    ) external pure returns (Outcome memory);
+    function reprioritize(Outcome calldata allocation, Outcome calldata guarantee)
+        external
+        pure
+        returns (Outcome memory);
 
-    function moveAuthorized(Commitment.CommitmentStruct calldata _commitment, Signature calldata signature) external pure returns (bool);
+    function moveAuthorized(
+        Commitment.CommitmentStruct calldata _commitment,
+        Signature calldata signature
+    ) external pure returns (bool);
 
-        function reduce(
-        Outcome memory outcome,
-        address recipient,
-        uint amount,
-        address token
-    ) public pure returns (Outcome memory);
+    function reduce(Outcome memory outcome, address recipient, uint256 amount, address token)
+        public
+        pure
+        returns (Outcome memory);
 }
 
 contract NitroAdjudicator {
     using Commitment for Commitment.CommitmentStruct;
-    using SafeMath for uint;
+    using SafeMath for uint256;
     INitroLibrary Library; // Abs
 
     constructor(address _NitroLibraryAddress) public {
@@ -60,7 +65,7 @@ contract NitroAdjudicator {
     }
 
     // TODO: Challenge duration should depend on the channel
-    uint constant CHALLENGE_DURATION = 5 minutes;
+    uint256 constant CHALLENGE_DURATION = 5 minutes;
 
     struct Authorization {
         // Prevents replay attacks:
@@ -70,7 +75,7 @@ contract NitroAdjudicator {
         // to send the transaction from, preventing any replay attack.
         address participant; // the account used to sign commitment transitions
         address destination; // either an account or a channel
-        uint amount;
+        uint256 amount;
         address sender; // the account used to sign transactions
     }
 
@@ -81,7 +86,7 @@ contract NitroAdjudicator {
         INitroLibrary.Signature ultimateSignature; // Abs
     }
 
-    mapping(address => mapping(address => uint)) public holdings;
+    mapping(address => mapping(address => uint256)) public holdings;
     mapping(address => INitroLibrary.Outcome) public outcomes; // Abs
     address private constant zeroAddress = address(0);
 
@@ -89,17 +94,21 @@ contract NitroAdjudicator {
     // ETH and Token Management
     // **************
 
-
-function deposit(address destination, uint expectedHeld,
- uint amount, address token) public payable {
-       if (token == zeroAddress) {
-        require(msg.value == amount, "Insufficient ETH for ETH deposit");
+    function deposit(address destination, uint256 expectedHeld, uint256 amount, address token)
+        public
+        payable
+    {
+        if (token == zeroAddress) {
+            require(msg.value == amount, "Insufficient ETH for ETH deposit");
         } else {
             IERC20 _token = IERC20(token);
-            require(_token.transferFrom(msg.sender,address(this),amount), 'Could not deposit ERC20s');
-            }
+            require(
+                _token.transferFrom(msg.sender, address(this), amount),
+                "Could not deposit ERC20s"
+            );
+        }
 
-        uint amountDeposited;
+        uint256 amountDeposited;
         // This protects against a directly funded channel being defunded due to chain re-orgs,
         // and allow a wallet implementation to ensure the safety of deposits.
         require(
@@ -122,42 +131,40 @@ function deposit(address destination, uint expectedHeld,
         if (amountDeposited < amount) {
             // refund whatever wasn't deposited.
             if (token == zeroAddress) {
-              msg.sender.transfer(amount - amountDeposited); // TODO use safeMath here
-          }
-            else {
+                msg.sender.transfer(amount - amountDeposited); // TODO use safeMath here
+            } else {
                 IERC20 _token = IERC20(token);
                 _token.transfer(msg.sender, amount - amountDeposited); // TODO use safeMath here
                 // TODO compute amountDeposited *before* calling into erc20 contract, so we only need 1 call not 2
-                }
+            }
         }
         emit Deposited(destination, amountDeposited, holdings[destination][token]);
     }
 
-    function transferAndWithdraw(address channel,
+    function transferAndWithdraw(
+        address channel,
         address participant,
         address payable destination,
-        uint amount,
+        uint256 amount,
         address token,
         uint8 _v,
         bytes32 _r,
         bytes32 _s
     ) public payable {
         transfer(channel, participant, amount, token);
-        withdraw(participant, destination, amount, token, _v, _r ,_s);
+        withdraw(participant, destination, amount, token, _v, _r, _s);
     }
 
-    function withdraw(address participant,
+    function withdraw(
+        address participant,
         address payable destination,
-        uint amount,
+        uint256 amount,
         address token,
         uint8 _v,
         bytes32 _r,
         bytes32 _s
     ) public payable {
-        require(
-            holdings[participant][token] >= amount,
-            "Withdraw: overdrawn"
-        );
+        require(holdings[participant][token] >= amount, "Withdraw: overdrawn");
         Authorization memory authorization = Authorization(
             participant,
             destination,
@@ -172,30 +179,28 @@ function deposit(address destination, uint expectedHeld,
 
         holdings[participant][token] = holdings[participant][token].sub(amount);
         // Decrease holdings before calling to token contract (protect against reentrancy)
-        if (token == zeroAddress) {destination.transfer(amount);}
-        else {
+        if (token == zeroAddress) {
+            destination.transfer(amount);
+        } else {
             IERC20 _token = IERC20(token);
-            _token.transfer(destination,amount);
-            }
+            _token.transfer(destination, amount);
+        }
 
     }
 
-
-    function transfer(address channel, address destination, uint amount, address token) public {
+    function transfer(address channel, address destination, uint256 amount, address token) public {
         require(
             outcomes[channel].challengeCommitment.guaranteedChannel == zeroAddress,
             "Transfer: channel must be a ledger channel"
         );
-        require(
-            outcomes[channel].finalizedAt <= now,
-            "Transfer: outcome must be final"
-        );
-        require(
-            outcomes[channel].finalizedAt > 0,
-            "Transfer: outcome must be present"
-        );
+        require(outcomes[channel].finalizedAt <= now, "Transfer: outcome must be final");
+        require(outcomes[channel].finalizedAt > 0, "Transfer: outcome must be present");
 
-        uint channelAffordsForDestination = Library.affords(destination, outcomes[channel], holdings[channel][token]);
+        uint256 channelAffordsForDestination = Library.affords(
+            destination,
+            outcomes[channel],
+            holdings[channel][token]
+        );
 
         require(
             amount <= channelAffordsForDestination,
@@ -206,40 +211,40 @@ function deposit(address destination, uint expectedHeld,
         holdings[channel][token] = holdings[channel][token] - amount;
     }
 
-
-    function concludeAndWithdraw(ConclusionProof memory proof,
+    function concludeAndWithdraw(
+        ConclusionProof memory proof,
         address participant,
         address payable destination,
-        uint amount,
+        uint256 amount,
         address token,
         uint8 _v,
         bytes32 _r,
         bytes32 _s
-    ) public{
+    ) public {
         address channelId = proof.penultimateCommitment.channelId();
-        if (outcomes[channelId].finalizedAt > now || outcomes[channelId].finalizedAt == 0){
-        _conclude(proof);
+        if (outcomes[channelId].finalizedAt > now || outcomes[channelId].finalizedAt == 0) {
+            _conclude(proof);
         } else {
-            require(keccak256(abi.encode(proof.penultimateCommitment)) == keccak256(abi.encode(outcomes[channelId].challengeCommitment)),
-            "concludeAndWithdraw: channel already concluded with a different proof");
+            require(
+                keccak256(abi.encode(proof.penultimateCommitment)) ==
+                    keccak256(abi.encode(outcomes[channelId].challengeCommitment)),
+                "concludeAndWithdraw: channel already concluded with a different proof"
+            );
         }
-        transfer(channelId,participant, amount, token);
-        withdraw(participant,destination, amount, token, _v,_r,_s);
+        transfer(channelId, participant, amount, token);
+        withdraw(participant, destination, amount, token, _v, _r, _s);
     }
 
-    function claim(address guarantor, address recipient, uint amount, address token) public {
+    function claim(address guarantor, address recipient, uint256 amount, address token) public {
         INitroLibrary.Outcome memory guarantee = outcomes[guarantor]; // Abs
         require(
             guarantee.challengeCommitment.guaranteedChannel != zeroAddress,
             "Claim: a guarantee channel is required"
         );
 
-        require(
-            isChannelClosed(guarantor),
-            "Claim: channel must be closed"
-        );
+        require(isChannelClosed(guarantor), "Claim: channel must be closed");
 
-        uint funding = holdings[guarantor][token];
+        uint256 funding = holdings[guarantor][token];
         INitroLibrary.Outcome memory reprioritizedOutcome = Library.reprioritize( // Abs
             outcomes[guarantee.challengeCommitment.guaranteedChannel],
             guarantee
@@ -254,11 +259,9 @@ function deposit(address destination, uint expectedHeld,
             holdings[guarantor][token] = holdings[guarantor][token].sub(amount);
             holdings[recipient][token] = holdings[recipient][token].add(amount);
         } else {
-            revert('Claim: guarantor must be sufficiently funded');
+            revert("Claim: guarantor must be sufficiently funded");
         }
     }
-
-
 
     // **********************
     // ForceMove Protocol API
@@ -268,17 +271,12 @@ function deposit(address destination, uint expectedHeld,
         _conclude(proof);
     }
 
-
-
     function forceMove(
         Commitment.CommitmentStruct memory agreedCommitment,
         Commitment.CommitmentStruct memory challengeCommitment,
         INitroLibrary.Signature[] memory signatures // Abs
     ) public {
-        require(
-            !isChannelClosed(agreedCommitment.channelId()),
-            "ForceMove: channel must be open"
-        );
+        require(!isChannelClosed(agreedCommitment.channelId()), "ForceMove: channel must be open");
         require(
             Library.moveAuthorized(agreedCommitment, signatures[0]),
             "ForceMove: agreedCommitment not authorized"
@@ -302,19 +300,16 @@ function deposit(address destination, uint expectedHeld,
             challengeCommitment.token
         );
 
-        emit ChallengeCreated(
-            channelId,
-            challengeCommitment,
-            now + CHALLENGE_DURATION
-        );
+        emit ChallengeCreated(channelId, challengeCommitment, now + CHALLENGE_DURATION);
     }
 
-    function refute(Commitment.CommitmentStruct memory refutationCommitment, INitroLibrary.Signature memory signature) public { // Abs
+    function refute(
+        Commitment.CommitmentStruct memory refutationCommitment,
+        INitroLibrary.Signature memory signature
+    ) public {
+        // Abs
         address channel = refutationCommitment.channelId();
-        require(
-            !isChannelClosed(channel),
-            "Refute: channel must be open"
-        );
+        require(!isChannelClosed(channel), "Refute: channel must be open");
 
         require(
             Library.moveAuthorized(refutationCommitment, signature),
@@ -322,7 +317,13 @@ function deposit(address destination, uint expectedHeld,
         );
 
         require(
-            Rules.validRefute(outcomes[channel].challengeCommitment, refutationCommitment, signature.v, signature.r, signature.s),
+            Rules.validRefute(
+                outcomes[channel].challengeCommitment,
+                refutationCommitment,
+                signature.v,
+                signature.r,
+                signature.s
+            ),
             "Refute: must be a valid refute"
         );
 
@@ -337,12 +338,13 @@ function deposit(address destination, uint expectedHeld,
         outcomes[channel] = updatedOutcome;
     }
 
-    function respondWithMove(Commitment.CommitmentStruct memory responseCommitment, INitroLibrary.Signature memory signature) public { // Abs
+    function respondWithMove(
+        Commitment.CommitmentStruct memory responseCommitment,
+        INitroLibrary.Signature memory signature
+    ) public {
+        // Abs
         address channel = responseCommitment.channelId();
-        require(
-            !isChannelClosed(channel),
-            "RespondWithMove: channel must be open"
-        );
+        require(!isChannelClosed(channel), "RespondWithMove: channel must be open");
 
         require(
             Library.moveAuthorized(responseCommitment, signature),
@@ -350,7 +352,13 @@ function deposit(address destination, uint expectedHeld,
         );
 
         require(
-            Rules.validRespondWithMove(outcomes[channel].challengeCommitment, responseCommitment, signature.v, signature.r, signature.s),
+            Rules.validRespondWithMove(
+                outcomes[channel].challengeCommitment,
+                responseCommitment,
+                signature.v,
+                signature.r,
+                signature.s
+            ),
             "RespondWithMove: must be a valid response"
         );
 
@@ -371,14 +379,9 @@ function deposit(address destination, uint expectedHeld,
         Commitment.CommitmentStruct memory _responseCommitment,
         INitroLibrary.Signature memory _alternativeSignature, // Abs
         INitroLibrary.Signature memory _responseSignature // Abs
-    )
-      public
-    {
+    ) public {
         address channel = _responseCommitment.channelId();
-        require(
-            !isChannelClosed(channel),
-            "AlternativeRespondWithMove: channel must be open"
-        );
+        require(!isChannelClosed(channel), "AlternativeRespondWithMove: channel must be open");
 
         require(
             Library.moveAuthorized(_responseCommitment, _responseSignature),
@@ -396,7 +399,6 @@ function deposit(address destination, uint expectedHeld,
         bytes32[] memory s = new bytes32[](2);
         s[0] = _alternativeSignature.s;
         s[1] = _responseSignature.s;
-
 
         require(
             Rules.validAlternativeRespondWithMove(
@@ -459,6 +461,12 @@ function deposit(address destination, uint expectedHeld,
     );
     event Concluded(address channelId);
     event Refuted(address channelId, Commitment.CommitmentStruct refutation);
-    event RespondedWithMove(address channelId, Commitment.CommitmentStruct response, uint8 v, bytes32 r, bytes32 ss);
+    event RespondedWithMove(
+        address channelId,
+        Commitment.CommitmentStruct response,
+        uint8 v,
+        bytes32 r,
+        bytes32 ss
+    );
     event RespondedWithAlternativeMove(Commitment.CommitmentStruct alternativeResponse);
 }
